[sailfish-locale] Import from settings-system. JB#27456
Import sources as of revision 1.1.65 of settings-system.

If you want to inspect the history of the imported content, fetch the
replace reference first with:

git config --add remote.origin.fetch "+refs/replace/*:refs/replace/*"
git fetch

(replace "origin" with the actual name of your remote)

After that, it will appear as a branch merged with this commit. See
git-replace(1) for more details.